refactor: handleChangeTask moved to TaskList.jsx
